+++ src/PSRule.Rules.Azure.BuildTool/ProviderResourceOption.cs
@@ -1,0 +1,9 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+namespace PSRule.Rules.Azure.BuildTool;
+
+internal sealed class ProviderResourceOption
+{
+    public string OutputPath { get; set; }
+}